// sourcecode/BTrak.Packages/snova-activity-tracker-module/activity-tracker-components/activity-tracker-components/src/lib/activitytracker/models/top-sites.model.ts
export class TopSitesModel {
    appUrlName: string
    appUrlImage: string
    applicationTypeName: string
    neutral: any
    productive: any
    unProductive: any
    isApp: boolean;
    mySelf: boolean;
    onDate: string;
    isProductiveApps: boolean;
    totalTimeInHr: any;
    userName: string;
    idleTime: any;
    totalTime: any;
    userId: string[];
    roleId: string[];
    branchId: string[];
    dateFrom: any;
    dateTo: any;
    isForSummary: boolean;
    productiveInSec: any;
    unProductiveInSec: any;
    neutralInSec: any;
}